Safari-609.1.11.1/LayoutTests/http/tests/websocket/tests/hybi/closed-when-entering-page-cache-expected.txt

CONSOLE MESSAGE: WebSocket connection to 'ws://127.0.0.1:8880/websocket/tests/hybi/echo' failed: WebSocket is closed due to suspension.
Test that an open WebSocket does not prevent a page from entering page cache.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


pageshow - not from cache
pagehide - entering cache
pageshow - from cache
PASS Page did enter and was restored from the page cache
PASS error event fired
PASS wasRestoredFromPageCache is true
PASS close event fired
PASS wasRestoredFromPageCache is true
PASS receivedErrorEvent is true
PASS closeEvent.wasClean is false
PASS closeEvent.code is 1006
PASS successfullyParsed is true

TEST COMPLETE
